[REF] move generate nfe key to l10n_br_nfe module
